add put endpoint

+++ index.js
@@ -7,6 +7,7 @@ const db = require("./data/db.js");
 // add your server code starting here
 const server = express();
 server.use(bodyParser.json());
+server.use(cors());
 
 server.get("/", (req, res) => {
   res.status(200).json({ message: "Welcome to my first API" });
@@ -26,7 +27,7 @@ server.post("/api/posts", (req, res) => {
   }
 });
 
-server.get("/api/posts", cors(),(req, res) => {
+server.get("/api/posts",(req, res) => {
   db.find()
     .then(posts => res.status(200).json(posts))
     .catch(error => res.status(500).json(error));
@@ -48,21 +49,31 @@ server.get("/api/posts/:id", (req, res) => {
     .catch(error => res.status(500).json(error));
 });
 
-server.delete("api/posts/:id", (req, res) => {
+server.put("api/posts/:id", (req, res) => {
+  const { title, contents } = req.body;
   const postId = req.params.id;
 
-  db.findById(postId)
-    .then(post => {
-      if (post.length === 0) {
-        db.delete(postId)
-          .then()
-      } else {
-        resolve(post)
-      }
-      
-    })
-    
+  if (!title && !contents) {
+    res.status(400).json({ message: 'Did not receive title or contents to update'})
+  } else {
 
-});
+    db.findById(postId)
+      .then(post => {
+        if (post.length === 0) {
+          res.status(404).json({ message: 'Cannot find a post with requested id' });
+        } else {
+          db.delete(postId)
+            .then(num => {
+              if (num === 1) {
+                res.status(200).json(post);
+              } else {
+                res.status(500).json({ message: 'Cannot perform update' })
+              }
+            })
+        }
+      })
+      .catch(error => res.status(500).json(error))
+  }
+})
 
 server.listen(5000, () => console.log("Server is running on port: 5000"));
